fix database connection
