+++ database.js
@@ -1,20 +1,12 @@
 var mysql = require('mysql');
 
 function Database() {
-    this.initial = () => {
-        this.pool = mysql.createPool({
-            host: 'localhost',
-            user: 'root',
-            password: '',
-            database: 'enviromentmonitoring'
-        });
-    };
-
-    this.acquire = (callback) => {
-        this.pool.getConnection((err, connection) => {
-            callback(err, connection);
-        });
-    };
+    this.connection = mysql.createConnection({
+        host: 'localhost',
+        user: 'root',
+        password: '',
+        database: 'environmentmonitoring'
+    }).connect();
 };
 
 module.exports = new Database();
